chore: move hooks
