+++ features/earn/vault-ggv/deposit/ggv-deposit-warning.tsx
@@ -1,6 +1,6 @@
 import { VaultWarning } from 'features/earn/shared/vault-warning';
 
-import { useGGVDepositStatus } from '../hooks/use-ggv-deposit-status';
+import { useGGVDepositStatus } from './hooks/use-ggv-deposit-status';
 
 const WARNING_TEXT = {
   cap: 'The vault has reached its deposit limit.',
