+++ features/earn/vault-ggv/deposit/hooks/use-ggv-deposit-status.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { useMainnetOnlyWagmi } from 'modules/web3';
-import { getGGVTellerContract, getGGVVaultContract } from '../contracts';
+import { getGGVTellerContract, getGGVVaultContract } from '../../contracts';
 import { INFINITE_DEPOSIT_CAP } from './use-ggv-max-deposit';
 
 type DepositStatusResponse = {
